requirements.txt
kafka-python==2.0.2
PyYAML==5.4.1
schema==0.7.5
confluent_kafka==1.9.2